// clean-arch/src/usecases/customer/update/update.customer.dto.ts
export interface InputUpdateCustomerDto {
  id: string;
  name: string;
  address: {
    street: string;
    number: number;
    zip: string;
    city: string;
  };
}

export interface OutputUpdateCustomerDto {
  id: string;
  name: string;
  address: {
    street?: string;
    number?: number;
    zip?: string;
    city?: string;
  };
}